MNT restructured utils; add model performance test function
