+++ deepnog/utils/__init__.py
@@ -1,5 +1,7 @@
 from .bio import EXTENDED_IUPAC_PROTEIN_ALPHABET, SeqIO, parse
-from .io_utils import create_df, get_config, get_data_home, get_logger, get_weights_path
+from .config import get_config
+from .io_utils import create_df, get_data_home, get_weights_path
+from .logger import get_logger
 from .sync import SynchronizedCounter
 from .network import count_parameters, load_nn, set_device
